Moving SearchBoundary to ArcgisUtil
